move event spanning handling to utils and memoize
